+++ _promise_a1.d.ts
@@ -1,2 +1,2 @@
 import type { rush_project_type } from './rush_project_type';
-export declare function _promise_a1(projects: rush_project_type[], _promise: any): any[];
+export declare function _promise_a1<O>(projects: rush_project_type[], _promise: (project: rush_project_type) => Promise<O>): Promise<O>[];