+++ _promise_a1.ts
@@ -1,6 +1,9 @@
 import type { rush_project_type } from './rush_project_type'
-export function _promise_a1(projects: rush_project_type[], _promise) {
-	const promise_a1 = []
+export function _promise_a1<O>(
+	projects:rush_project_type[],
+	_promise:(project:rush_project_type) => Promise<O>
+) {
+	const promise_a1 = [] as Promise<O>[]
 	for (let i = 0; i < projects.length; i++) {
 		const project = projects[i]
 		promise_a1.push(_promise(project))