+++ bin/pwd--packages.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+TEMP=$(mktemp); cat <<EOF > $TEMP; pnpm recursive exec -- sh "$TEMP"
+pwd
+EOF