+++ bin/tsc-clean--packages.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+TEMP=$(mktemp); cat <<EOF > $TEMP; pnpm recursive exec -- sh "$TEMP"
+rm -f tsconfig.tsbuildinfo
+EOF