+++ map_package_json_path_glob.d.ts
@@ -1,1 +1,1 @@
-export declare function map_package_json_path_glob(glob_txt: any, fn: any): Promise<any[]>;
+export declare function map_package_json_path_glob(glob_txt: any, fn: any): Promise<unknown[]>;